rename circuit_transpiler module to digital
