+++ docs/code/transpiler.rst
@@ -1,4 +1,4 @@
 ql.transpiler
 ===============
 
-.. automodule:: qililab.circuit_transpiler
+.. automodule:: qililab.digital
